+++ src/mockingcase.d.ts
